Use update_provider_tree in vmware virt driver
Replace get_inventory with update_provider_tree, using most of the
same code that get_inventory used.

As yet, this does not support reshaping.

A unit test for get_inventory is updated to use update_provider_tree.
A provider tree is established, checked to confirm it has no inventory,
then updated and new inventory confirmed. This appears to have been the
one unit test that touched get_inventory in the vmware virt driver so
much of the real testing of this will be when the 3rd party CI hits it.
(More robust testing will be added when reshaping, shared disk support,
and tenant VDC are added.)

Some TODOs are added in the update_provider_tree implementation to
indicate differences from the libvirt implementation and the spots
where new functionality will be added in future patches.

Change-Id: Ib9bd7617e36e28a29e2d15c0110de88b375030d1