Add archive_deleted_rows wrinkle to cross-cell functional test
This adds a wrinkle to the test_resize_confirm_image_backed
functional test which will run the archive_deleted_rows DB API
method on all cells (there are three in the test: cell0, cell1 and
cell2) and verify at least one instance is archived. This is mostly
a sanity check to make sure some referential constraint is not
messed up during the cross-cell resize + confirm flow where the
source cell1 instance records are hard-deleted and the target
cell2 instance records are soft-deleted.

Part of blueprint cross-cell-resize

Change-Id: I444a9f94060dd620a4837ba123544c1fb0fea190