Func test for migrate server with ports having resource request
A new functional test proves that bandwidth allocation is made on the
destination host during the migration and the binding:profile of the
port is updated to point to the destination resource provider.

As a result, migrate_instance_finish has to be un-stubbed in the
NeutronFixture to update the port's binding profile.

blueprint: support-move-ops-with-qos-ports

Change-Id: Ia69dace6d3b395fa113c1382e8eb52c2cc36119d