Add BFV wrinkle to TestNovaManagePlacementHealAllocations
The heal_allocations command is using the
nova.scheduler.utils.resources_from_flavor method to build
the dictionary of resource allocations to PUT to placement
and that method has a volume-backed instance check in it to
avoid creating DISK_GB allocations for volume-backed servers.
We weren't testing any volume-backed servers in the
heal_allocations functional tests though so this adds a
wrinkle for that coverage.

Change-Id: I3847567601a5bd38e9774f75914b6dda50c9dd40