Move _fill_provider_mapping to the scheduler_utils
The _fill_provider_mapping is called during boot and reschedule so far
but to support move operations with ports having resource request we
need to call it during various coductor tasks. So this patch move the
function to the utils and adjust the name and the signature accordingly.

blueprint: support-move-ops-with-qos-ports

Change-Id: I76f777e4f354b92c55dbd52a20039e504434b3a1