Add regression test for bug 1825034
The fill_virtual_interface_list online data migration creates a
fake mostly empty instance record to satisfy a foreign key constraint
in the virtual_interfaces table which is used as a marker when paging
across cells to fulfill the migration. The problem is if you list deleted
servers (as admin) with the all_tenants=1 and deleted=1 filters, the API
will fail with a 500 error trying to load the instance.flavor field.

This adds a functional regression test for the bug.

Change-Id: I2030412566dfc6ec23dbf37685f6e6d145f710dc
Related-Bug: #1825034