Add functional regression recreate test for bug 1825020
Change I06fad233006c7bab14749a51ffa226c3801f951b in Stein
started validating extra specs on the new flavor during a
resize but didn't take into account some validation that
happens with disk size with the new flavor if the validation
code isn't told that the server is volume-backed. That results
in a 500 FlavorDiskSmallerThanMinDisk response from the API
if the new flavor has a smaller disk size than the current
flavor that the server is using, which shouldn't actually
matter for a volume-backed server.

This adds a functional regression test to recreate the bug.

Change-Id: I9e7e44727946705506d740896349ca1833cfddcd
Related-Bug: #1825020