Fix {min|max}_version in ironic Adapter setup
Change If625411f40be0ba642baeb02950f568f43673655 introduced
nova.utils.get_ksa_adapter, which accepts min_version and max_version
kwargs to be passed through to the ksa Adapter constructor. These are
supposed to represent minimum and maximum *major* API versions,
min_version was erroneously set to *microversions* when setting up the
Adapter for ironicclient. This commit changes it to a major version.
(Microversion negotiation is done within ironicclient itself.)

Also, this bug went latent for several releases because a) it only seems
to be triggered when region_name is given in the conf; but also b)
ironicclient has code to discover a reasonable endpoint if passed None.
So this change also adds a warning log if we try and fail to discover
the endpoint via ksa.

Conflicts:
	nova/tests/unit/virt/ironic/test_client_wrapper.py
This was just because the old microversion was 1.38 instead of 1.46. The
patch still changes it to 1.0.

Change-Id: I34a3f8d4a496217eb01790e2d124111625bf5f85
Closes-Bug: #1825583
(cherry picked from commit 13278be9f265e237fc68ee60acfacaa1df68522e)
(cherry picked from commit e6ca3838d940dea417cdeffe22c656d222737a7e)