Fix {min|max}_version in ironic Adapter setup
Change If625411f40be0ba642baeb02950f568f43673655 introduced
nova.utils.get_ksa_adapter, which accepts min_version and max_version
kwargs to be passed through to the ksa Adapter constructor. These are
supposed to represent minimum and maximum *major* API versions,
min_version was erroneously set to *microversions* when setting up the
Adapter for ironicclient. This commit changes it to a major version.
(Microversion negotiation is done within ironicclient itself.)

Also, this bug went latent for several releases because a) it only seems
to be triggered when region_name is given in the conf; but also b)
ironicclient has code to discover a reasonable endpoint if passed None.
So this change also adds a warning log if we try and fail to discover
the endpoint via ksa.

Change-Id: I34a3f8d4a496217eb01790e2d124111625bf5f85
Closes-Bug: #1825583
(cherry picked from commit 13278be9f265e237fc68ee60acfacaa1df68522e)