Implement update_provider_tree
For now, we'll just set the available ram, disk and vcpus.
This will allow allocation ratios to be set at the placement
inventory level. At the same time, driver capabilities will be
exposed through traits.

Once #1784020 is addressed, we may skip reporting shared storage
capacity.

We're also skipping CPU features for a couple of reasons:
* there are just a few features defined by os-traits, yet even
  fewer exposed by Windows APIs. One option would be to call cpuid
  to get a more exhaustive list of features but that would require
  a C Python extension or separate executable.
* we can't tell for sure which CPU features will actually be exposed
  to the guest, especially when limiting guest CPU features in order
  to support live migration in heterogeneous deployments.

NOTE(mriedem): This is a cherry-pick of the out-of-tree hyperv
driver change https://review.opendev.org/655123/. The only
modification is the os_resource_classes imports are grouped with
third party imports rather than the nova imports.

Change-Id: Ibfb879751bba038054ba7852dc49f6d9e67027d6