re-calculate provider mapping during migration
Nova intentionally does not persist the resoruce request of the neturon
ports. Therefore during migration nova needs to query neturon about the
resource requests to include them to the allocation_candidates query
sent to placement during scheduling. Also when the allocation is made by
the scheduler nova needs to re-calculate request group - resource
provider mapping. A subsequent patch will use this mapping to update the
binding profile when the port is bound to the destination host of the
migration.

blueprint: support-move-ops-with-qos-ports

Change-Id: I8e5a0480c81ba548bc1f50a8098eabac52b11453