Add nova-multi-cell job
This adds a non-voting nova-multi-cell job which
is a 2-node job with 2 non-cell0 cells (cell1 and cell2).
The compute on the controller will be in cell1 and the
subnode compute will be in cell2.

Until cross-cell resize is supported, cold migration
has to be disabled since we can't cold migrate to
the same host (and there is only one compute host per
cell). As a result, resize to the same host is enabled
(it is normally disabled in the tempest multinode job).
The parent job enables live migration testing but since
we only have one host per cell and don't support live
migration across cells, the live migration tests are
also disabled.

The job is also setup to run with python3 since that
is the way of the future.

Non-essential services, like cinder-backup, are disabled
since we (1) won't need it for the tests we run and
(2) would like to save as much room on the hosts as possible
since we're running multiple cell conductors on the
controller host.

Since this is new and its already uncovered at least one
bug (bug 1826472), the job is non-voting in the check queue
for now until we're happy with its stability.

Related to blueprint cross-cell-resize

Change-Id: I8ad3f8bcbecf0f6b1d0f57e9f61ff21bd7fbe365