pull out put_allocation call from _heal_*
Both allocation healing steps calls the placement API. This patch pulls
out the placement updating code to a single place. To do that it change
the healing steps to only generate / update the allocation individually
and then at the end of the healing there will be a single placement
update with this allocation.

This will help us to include the port related allocation into the instance
allocation by modifying a single place in the code.

Related-Bug: #1819923

Change-Id: I0e9f9a488141da599c10af8cabb4f6a5d111104f