Add request_spec to server move RPC calls
To be able to fill the allocation key in the port binding:profile during
the move operations the nova-compute needs to get the RequestSpec object
to have access to the port - resource provider mapping calculated in the
conductor.

This patch bumps the compute RPC api version and adds a new request_spec
parameter to multiple calls. Also it makes sure that the request_spec is
passed by the sender.

Change-Id: If1f465112b8e9b0304b8b5b864b985f72168d839
blueprint: support-move-ops-with-qos-ports