update allocation in binding profile during migrate
If the server has port with resource allocation and the server is
migrated then when the port is bound to the destination host the
allocation key needs to be updated in the binding:profile to point to
the resource provider that provides resources for this port on the
destination host.

This patch extends the migrate_instance_finish() network api method to
pass the updated resource providers of the ports during migration.

Change-Id: I220fa02ee916728e241503084b14984bab4b0c3b
blueprint: support-move-ops-with-qos-ports