Find instance in another cell during floating IP re-association
When associating a floating IP to instance A but it was already
associated with instance B, we try to refresh the info cache on
instance B. The problem is the context is targeted to the cell
for instance A and instance B might be in another cell, so we'll
get an InstanceNotFound error trying to lookup instance B.

This change tries to find the instance in another cell using its
instance mapping, and makes the code a bit more graceful if
instance B is deleted.

Conflicts:
      devstack/nova-multi-cell-blacklist.txt

NOTE(mriedem): The conflict is because the nova-multi-cell
job does not exist in Stein.

Change-Id: I71790afd0784d98050ccd7cc0e046321da249cbe
Closes-Bug: #1826472
(cherry picked from commit 481cb5ce04e220c26e5772f4253d63e212adca45)