Remove global state from the FakeDriver
The virt driver FakeDriver used in both the functional and in the unit
test used a global state to configure the host and node names the driver
reports. This was hard to use when more then one compute service is started.
Also global state is dangerous.

It turned out that only a set of unit tests are using multiple nodes per
compute the rest of the tests can simply use host=<hostname>,
nodes=[<hostname>] setup.

So this removes the global state.

On stable/stein:
* Changes in functional/regressions/test_bug_1764556.py just removes the
  usage of the global state as that test was backported before the patch
  being backported was introduced

* Changes in functional/regressions/test_bug_1839560.py is a re-backport
  of If822509e906d5094f13a8700b2b9ed3c40580431 adapted to the current
  patch.

Related-Bug: #1859766

Change-Id: I2cf2fcbaebc706f897ce5dfbff47d32117064f9c
(cherry picked from commit b5666fb49239dbed6777481be63276db35ed8118)