Avoid unnecessary joins in delete_resource_provider
If cascade=True, we're getting all of the instances on the
compute node just to use the uuid, which will by default
join on the info_cache and security_groups for the instances.
This is a simple optimization to avoid those unnecessary joins.

A TODO is left to further optimize this with a new InstanceList
query method to just get the instance uuids on a given host/node
combo, but that requires an RPC change which we can't backport.

Change-Id: Ie121210456a240c257979d3269db115ddae2d23b
Related-Bug: #1811726
Related-Bug: #1756179