Add a workaround config toggle to refuse ceph image upload
If a compute node is backed by ceph, and the image is not clone-able
in that same ceph, nova will try to download the image from glance
and upload it to ceph itself. This is nice in that it "just works",
but it also means we store that image in ceph in an extremely
inefficient way. In a glance multi-store case with multiple ceph
clusters, the user is currently required to make sure that the image
they are going to use is stored in a backend local to the compute
node they land on, and if they do not (or can not), then nova will
do this non-COW inefficient copy of the image, which is likely not
what the operator expects.

Per the discussion at the Denver PTG, this adds a workaround flag
which allows the operators to direct nova to *not* do this behavior
and instead refuse to boot the instance entirely.

Conflicts:
    nova/conf/workarounds.py

NOTE(melwitt): The conflict is because this patch originally landed on
ussuri and change If874f018ea996587e178219569c2903c2ee923cf (Reserve
DISK_GB resource for the image cache) landed afterward and was
backported to stable/train.

Related-Bug: #1858877
Change-Id: I069b6b1d28eaf1eee5c7fb8d0fdef9c0c229a1bf
(cherry picked from commit 80191e6d828cf823ce3aa7c6176da5e531694900)