Optimize SchedulerReportClient.delete_resource_provider
This is used to optimize the SchedulerReportClient
delete_resource_provider method when deleting a compute
node and its related resource provider which happens
in both the API when a compute service is deleted and
in the compute service in the update_available_resource
method for orphan nodes (so the new InstanceList query
method needs to be remotable).

The actual DB query stuff in here is tested in the
nova.tests.functional.wsgi.test_services code.

Change-Id: Id033e8f95f853ccfc34bd350be6df11c1bfb1b7d
Related-Bug: #1811726