Log when port resource is leaked during port delete
When a bound port is deleted in neutron, nova gets a
network-vif-deleted notification and the port is
detached from the server. However if the port had resource request
then the resulting allocation is leaked.

This patch makes sure nova at least logs an ERROR. Also this patch
asserts that the leaked allocation is reclaimed when the server is
deleted.

Change-Id: I5d905aeb5b25f84d406dbf238d0d3a46f0f81161
Related-Bug: #1820588