Make nova-tox-functional-py36 reusable
We'd like placement to be able to run this job itself. Instead of
duplicating it, we can simply use it directly. However, according
to discussions with the infra gang, to ensure that works properly
we need to make openstack/nova a required project and explicitly
set zuul_work_dir. These are the default values already when
used from nova itself.

Change-Id: I73f9dacd912331d209a6fa04eaf5591d94d720f2
Story: 2005562