[Docs] Change the server query parameter display into a list.
The server query parameters that are allowed for users are displayed
in a single paragraph which is difficult to parse and some of the
parameters are missing the microversions they are available from.
This problem exists in the server_concepts api-guide documentation
while its already in the list form in the servers api-reference
documentation. This patch basically copies the list over from the
api-guide.

Change-Id: I27c3d1882ab904933129bd3414ccded8c3c56b1e