Add extra logging to request filters
This adds a timing decorator to request_filter.py which logs the
filters that we ran and how long they took. This requires a small
calling convention change for these request filters to return
True if they are enabled and/or did anything.

This also sprinkles some debug logging into all the existing filters
to provide contextual information about what they are doing to
make debugging scheduling decisions easier.

Change-Id: Ibf62c5398a95c75fde815309e7a4f82f1bd9649c