Robustify attachment tracking in CinderFixtureNewAttachFlow
Cinder allows more than one attachment record on a non-multiattach
volume as long as only one attachment is "connected" (has a host
connector).

When you create an empty (no connector) attachment for an in-use
volume, the volume status changes to "attaching". If you try to
update the empty attachment before deleting the previous attachment,
Cinder will return a 400 like:

  Invalid volume: Volume b2aba195-6570-40c4-82bb-46d3557fceeb status
  must be available or downloading to reserve, but the current status
  is attaching.

This change refactors the attachment tracking in the
CinderFixtureNewAttachFlow fixture to be able to track the connector
per attachment and if code tries to update an attachment on an
already connected volume it will fail.

Change-Id: I369f82245465e96fc15d4fc71a79a8a71f6f2c6d