Stop logging traceback when skipping quiesce
During snapshot of a volume-backed instance, we attempt to quiesce the
instance before doing the snapshot. If quiesce is not supported or the
qemu guest agent is not enabled, we will skip the quiesce and move on
to the snapshot.

Because quiesce is a call to nova-compute over RPC, when the libvirt
driver raises QemuGuestAgentNotEnabled, oslo.messaging will append the
full traceback to the exception message [1] for the remote caller. So,
a LOG.info(..., exp) log of the exception object will result in a log
of the full traceback. Logging of the full traceback causes confusion
for those debugging CI failures.

We would rather not log the full traceback in this case where we are
catching the exception and emitting an INFO message, so we should use
exp.format_message() instead of oslo.messaging's __str__ override.

[1] https://github.com/openstack/oslo.messaging/blob/40c25c2/oslo_messaging/_drivers/common.py#L212

Related-Bug: #1824315

Change-Id: Ibfedcb8814437c53081f5a2993ab84b25d73e557
(cherry picked from commit 66070416ab670daf4734f2c30ad2df5d96e615e6)
(cherry picked from commit 2f7b1031a9b80d15052fcd9b22831f1db0486a4a)