Use SDK for add/remove instance info from node
We would like nova not to use ironicclient, but instead to invoke the
ironic API directly through an openstacksdk connection.

The parent commits set up the framework, and this commit uses it
for adding and removing instance info from the node.

Depends on field mapping in the OpenStack SDK. That work maps SDK
field names to Ironic API field names within the SDK which allows
for consistency between fields in calls and parameters on returned
objects.

Change-Id: Id427e7923ff3a9d2957586fba5ccef0216318e6f