Support old & new versions of svm and vmx traits
I1c9a72d19ef9dadfb931efa3894867099974bcc7 added new Intel- and
AMD-specific trait namespaces for certain X86 CPU features, including a
couple that had already been defined in the generic X86 namespace (the
Intel-specific 'vmx' feature and the AMD-specific 'svm' feature).
Unfortunately, the libvirt driver was already exposing the generic
versions of these. For backward compatibility reasons, we have to
continue exposing those versions of those traits. But we also need to
expose the new arch-namespaced versions.

We accomplish this by reworking libvirt's CPU_TRAITS_MAPPING such that
the values can be tuples of traits, all of which are set when the
feature corresponding to the key is enabled.

Change-Id: Ib29c975d3d4e1adf14e215b1a28d30681ee16a77